+++ data/ttcutprojectdata.h
@@ -68,7 +68,7 @@ class TTCutProjectData
     QDomElement writeAudioSection(QDomElement& parent, const QString& filePath, int order, const QString& language, int delayMs = 0);
     QDomElement writeCutSection(QDomElement& parent, int cutIn, int cutOut, int order);
     QDomElement writeMarkerSection(QDomElement& parent, int markerPos, int markerType, int order);
-    QDomElement writeSubtitleSection(QDomElement& parent, const QString& filePath, int order, const QString& language);
+    QDomElement writeSubtitleSection(QDomElement& parent, const QString& filePath, int order, const QString& language, int delayMs = 0);
     void        parseVideoSection(QDomNodeList videoNodesList, TTAVData* avData);
     void        parseAudioSection(QDomNodeList audioNodesList, TTAVData* avData, TTAVItem* avItem);
     void        parseCutSection(QDomNodeList cutNodesList, TTAVItem* avItem);